Move collaborators into its own module
